+++ Assets/Scripts/Animations/CrystalAnimation.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using DG.Tweening;
+﻿using DG.Tweening;
 using UnityEngine;
 
 namespace Animations {
